Use the major/minor version of HHVM as API version in hphpize
We are currently relying on an untrusted and admittedly outdated value
of HHVM_API_VERSION for determining which directory any dynamic
extension will be installed in.

Since substantially there is no guarantee that extensions can be valid
across minor versions, using "MAJOR.MINOR" as an api version in hphpize
as a compromise. It will also allow system administrators to maintain a
stable link to the current extensions path if needed.