Default DynamicExtensionsPath to the value set by hphpize
Extensions built with hphpize will end up installed into a predefined
directory; it would make sense for HHVM to assume that is the default
value for the extensions to be located in, instead of defaulting to the
current working directory. This is by the way substantially what PHP
does with its extensions.