Use hphpize default as default for hhvm.dynamic_extensions_path
There is no reason for defaulting to the current working directory for
this setting. Given that any dynamic extension built with `hphpize` will
be configured and installed in a specific directory, default to that
specific directory; this will also make life easier for system
administrators during upgrades, as they don't need to change both
binaries AND config files at the same time.